requirements.txt
PyAutoIt==0.6.0
